# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gregg + Landon Current Client Cockpit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: clientId
          in: query
          required: false
          schema: { type: string }
        - name: owner
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /opportunity-signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List opportunity signals
      parameters:
        - name: clientId
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpportunitySignal"
  /opportunity-signals/{signalId}:
    parameters:
      - name: signalId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateSignal
      tags: [signals]
      summary: Update an opportunity signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunitySignal"
  /escalations:
    get:
      operationId: listEscalations
      tags: [escalations]
      summary: List escalations
      parameters:
        - name: clientId
          in: query
          required: false
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Escalations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escalation"
  /escalations/{escalationId}:
    parameters:
      - name: escalationId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateEscalation
      tags: [escalations]
      summary: Update an escalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        sourceCallNoteId:
          type: [string, "null"]
        title: { type: string }
        owner: { type: string }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        escalationFlag: { type: boolean }
        notes: { type: string }
      required:
        - id
        - clientId
        - sourceCallNoteId
        - title
        - owner
        - dueDate
        - priority
        - status
        - escalationFlag
        - notes
    TaskInput:
      type: object
      properties:
        clientId: { type: string }
        title: { type: string }
        owner: { type: string }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        escalationFlag: { type: boolean }
        notes: { type: string }
      required: [clientId, title]
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        owner: { type: string }
        dueDate: { type: string }
        priority: { type: string }
        status: { type: string }
        escalationFlag: { type: boolean }
        notes: { type: string }
    OpportunitySignal:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        sourceCallNoteId:
          type: [string, "null"]
        type: { type: string }
        description: { type: string }
        status: { type: string }
        routedTo: { type: string }
        createdAt: { type: string }
      required:
        - id
        - clientId
        - sourceCallNoteId
        - type
        - description
        - status
        - routedTo
        - createdAt
    SignalUpdate:
      type: object
      properties:
        type: { type: string }
        description: { type: string }
        status: { type: string }
        routedTo: { type: string }
    Escalation:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        sourceCallNoteId:
          type: [string, "null"]
        reason: { type: string }
        riskLevel: { type: string }
        routedTo: { type: string }
        decisionNeeded: { type: string }
        deadline: { type: string }
        status: { type: string }
      required:
        - id
        - clientId
        - sourceCallNoteId
        - reason
        - riskLevel
        - routedTo
        - decisionNeeded
        - deadline
        - status
    EscalationUpdate:
      type: object
      properties:
        reason: { type: string }
        riskLevel: { type: string }
        routedTo: { type: string }
        decisionNeeded: { type: string }
        deadline: { type: string }
        status: { type: string }
...